+++ src/core/Screen.cpp
@@ -319,7 +319,7 @@ void Screen::displayMonitorError()
 
 namespace {
     void deltaV() {
-        int x = AnalogInputs::getRealValue(AnalogInputs::deltaVout);
+        int16_t x = AnalogInputs::getRealValue(AnalogInputs::deltaVout);
         lcdPrintSigned(x, 5);
         lcdPrintChar('m');
         lcdPrintChar('V');
@@ -328,7 +328,7 @@ namespace {
     }
     void deltaT() {
         if(settings.externT_) {
-            int x = AnalogInputs::getRealValue(AnalogInputs::deltaTextern);
+        	int16_t x = AnalogInputs::getRealValue(AnalogInputs::deltaTextern);
             lcdPrintSigned(x*10, 5);
             lcdPrintChar('m');
             lcdPrintChar('C');
